requirements.txt
selenium==3.141.0
pytest==6.2.4
allure-pytest==2.8.40
webdriver-manager==3.5.2
PyAutoGUI==0.9.53
